pyproject.toml
[project]
name = "ai-quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.3",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
